telegram-bot/requirements.txt
python-telegram-bot
python-dotenv
fastapi
uvicorn
db-sqlite3